pyproject.toml
[project]
name = "b2-meister"
version = "0.1.0"
description = "Personal German B2 study companion — practice, test, HW-check"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.37",
    "agno>=0.5",
    "groq>=0.11",
    "ollama>=0.3",
    "openai>=1.40",
    "pymupdf>=1.24",
    "pypdf>=4.0",
    "psycopg[binary]>=3.2",
    "pgvector>=0.3",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = ["ruff", "ipython"]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
